requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
liboqs-python @ file:///home/keshav/Documents/quantsafe/pqc/liboqs-python
MarkupSafe==3.0.2
python-dotenv==1.1.1
Werkzeug==3.1.3
